Port core to Linux/GCC: headers, MSVC-only constructs, conformance
The game library never saw a conforming compiler. This is the bulk of
what stood between it and a GCC build:

Header hygiene. Server/Json/features.h shadowed glibc's <features.h> for
every translation unit in the game target, which broke libstdc++ before
any project code was even reached; renamed to json_features.h as
upstream jsoncpp did. Many headers relied on MSVC pulling in transitive
includes: added Define.h, ObjectGuid.h, Common.h, Position.h and the
standard containers where they are used, and forward declarations for
Aura, SpellValue, MoveSpline, Quest, VendorItem. HashFuctor.h
instantiates std::hash<ObjectGuid> at namespace scope and therefore
needs the specialisation, not a forward declaration.

Conformance. MSVC binds non-const references to temporaries; 280 local
declarations relied on that and now take copies, and the handful of
functions that only read their argument take it by const reference.
std::sinf and friends are not standard, std::sin is overloaded for
float.

Platform. Replaced itoa and sprintf_s with snprintf, dropped an unused
windows.h include from PathfindingMgr, guarded the MSVC-only
execution_character_set pragma and corecrt_math_defines.h, and
normalised 17 backslash include paths. Three includes differed from
their file only in case, which Windows hides and Linux does not.

+++ src/server/game/PlayerBot/FieldBotMgr.cpp
@@ -518,8 +518,8 @@ void FieldWarfare::TeleportToTargetVisible(Player* pBot, Player* pTarget)
 			break;
 		}
 		float targetAngle = frand(0, float(M_PI) * 2.0f);
-		distX = pTarget->GetPositionX() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.7f) * std::cosf(targetAngle);
-		distY = pTarget->GetPositionY() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.7f) * std::sinf(targetAngle);
+		distX = pTarget->GetPositionX() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.7f) * std::cos(targetAngle);
+		distY = pTarget->GetPositionY() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.7f) * std::sin(targetAngle);
 		distZ = pTarget->GetPositionZ();
 		distZ = pTarget->GetMap()->GetHeight(pTarget->GetPhaseMask(), distX, distY, distZ);
 	} while (!MapManager::IsValidMapCoord(pTarget->GetMapId(), distX, distY, distZ, pTarget->GetOrientation()));
@@ -554,8 +554,8 @@ Position FieldWarfare::GetEvadePosition(Player* centerPlayer, Player* evadePlaye
 	float maxDist = 0.0f;
 	for (float angle = 0.0f; angle < (float(M_PI) * 2.0f); angle += onceAngle)
 	{
-		float distX = evadePlayer->GetPositionX() + (BOTAI_FIELDTELEPORT_DISTANCE * 4) * std::cosf(angle);
-		float distY = evadePlayer->GetPositionY() + (BOTAI_FIELDTELEPORT_DISTANCE * 4) * std::sinf(angle);
+		float distX = evadePlayer->GetPositionX() + (BOTAI_FIELDTELEPORT_DISTANCE * 4) * std::cos(angle);
+		float distY = evadePlayer->GetPositionY() + (BOTAI_FIELDTELEPORT_DISTANCE * 4) * std::sin(angle);
 		float distZ = evadePlayer->GetPositionZ();
 		distZ = evadePlayer->GetMap()->GetHeight(evadePlayer->GetPhaseMask(), distX, distY, distZ);
 		Position pos(distX, distY, distZ, angle);
@@ -795,8 +795,8 @@ void FieldBotMgr::TeleportToTargetVisible(BotFieldAI* pFieldAI, Player* pTarget)
 				break;
 			}
 			float targetAngle = frand(0, float(M_PI) * 2.0f);
-			distX = pTarget->GetPositionX() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.2f) * std::cosf(targetAngle);
-			distY = pTarget->GetPositionY() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.2f) * std::sinf(targetAngle);
+			distX = pTarget->GetPositionX() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.2f) * std::cos(targetAngle);
+			distY = pTarget->GetPositionY() + (BOTAI_FIELDTELEPORT_DISTANCE * 1.2f) * std::sin(targetAngle);
 			distZ = pTarget->GetPositionZ();
 			distZ = pTarget->GetMap()->GetHeight(pTarget->GetPhaseMask(), distX, distY, distZ);
 		} while (!MapManager::IsValidMapCoord(pTarget->GetMapId(), distX, distY, distZ, pTarget->GetOrientation()) ||
